refactor(WIP): separate profile form and consent agreement
extracting the registration process into 2 pages with the second
being used for editing profile fields afterwards

TODO: create the profile page with the existing form, add redirects

[no ci]
