+++ client/src/views/Profile.vue
@@ -1,0 +1,10 @@
+<template>
+  <p>tmp</p>
+</template>
+
+<script lang="ts">
+import { Component, Vue } from "vue-property-decorator";
+
+@Component({})
+export default class App extends Vue {}
+</script>
