+++ server/src/services/account.ts
@@ -1,9 +1,11 @@
 import { ChatReport, User } from "@port-of-mars/server/entity";
 import { MoreThan, IsNull, Not, In, Repository, UpdateResult } from "typeorm";
+import validator from "validator";
 import { settings } from "@port-of-mars/server/settings";
 import { BaseService } from "@port-of-mars/server/services/db";
-import { generateUsername } from "@port-of-mars/server/util";
+import { generateUsername, ServerError } from "@port-of-mars/server/util";
 import { ModerationActionType, BAN, MUTE } from "@port-of-mars/shared/types";
+import _ from "lodash";
 
 const logger = settings.logging.getLogger(__filename);
 
@@ -165,6 +167,94 @@ export class AccountService extends BaseService {
     return await this.getRepository().update(id, { lastPlayerIp: ip });
   }
 
+  async updateProfile(user: User, data: { username: string; email: string; name: string }) {
+    const repo = this.em.getRepository(User);
+    let shouldVerifyEmail = !user.isVerified;
+    user.username = data.username;
+    user.name = data.name;
+    if (user.email && data.email && user.email !== data.email) {
+      user.email = data.email;
+      shouldVerifyEmail = true;
+    }
+    user.dateConsented = new Date();
+    // FIXME: consider consolidating validation logic in
+    // registration route (isEmailAvailable / isUsernameAvailable)
+    // here as well
+    if (!validator.isEmail(data.email)) {
+      throw new ServerError({
+        code: 400,
+        message: `Invalid email address`,
+        displayMessage: `Please enter a valid email address.`,
+      });
+    }
+    if (!validator.isLength(user.username, { min: 1, max: 30 })) {
+      throw new ServerError({
+        code: 400,
+        message: `Invalid username length`,
+        displayMessage: `Please choose a username between 1 and 30 characters long`,
+      });
+    }
+    if (!validator.isAlphanumeric(user.username, "en-US", { ignore: "_" })) {
+      throw new ServerError({
+        code: 400,
+        message: `Invalid username characters`,
+        displayMessage: `Please choose a username with only letters, numbers and underscores`,
+      });
+    }
+    await repo.save(user);
+    logger.debug("updated registration metadata for user %o", data);
+    if (shouldVerifyEmail) {
+      await this.sendEmailVerification(user);
+    }
+  }
+
+  createVerificationUrl(registrationToken: string) {
+    // FIXME: depends on VueRouter hash mode
+    return `${settings.host}/#/verify/${registrationToken}`;
+  }
+
+  async sendEmailVerification(u: User): Promise<void> {
+    if (_.isEmpty(u.email)) {
+      logger.warn("Trying to send email verification to a user with no email.");
+      return;
+    }
+    const verificationUrl = this.createVerificationUrl(u.registrationToken);
+    settings.emailer.sendMail(
+      {
+        from: `Port of Mars <${settings.supportEmail}>`,
+        to: u.email,
+        bcc: settings.supportEmail,
+        subject: "[Port of Mars] Verify your email",
+        // FIXME: convert these to a server-side template
+        text: `Greetings! Someone signed this email address up (${u.email}) to participate in the Port of Mars.
+      If this was you, please complete your registration by going to ${verificationUrl} and clicking on the "Verify" button.
+      If you did not request this, there is no need to take any action and you can safely ignore this message.
+
+      Thanks for participating!
+      the Port of Mars team`,
+        html: `Greetings!
+      <p>Someone signed this email address up (${u.email}) to participate in the Port of Mars.
+      If this was you, please complete your registration by <a href='${verificationUrl}'>going to ${verificationUrl}</a> 
+      and clicking on the "Verify" button.</p>
+      <p>If you did not request this, there is no need to take any action and you can safely ignore this message.</p>
+
+      <p>
+      Thanks for participating!
+      </p>
+      the Port of Mars team`,
+      },
+      function (err, info) {
+        if (err) {
+          logger.warn(`error : $err`);
+          throw new ServerError(err);
+        } else {
+          logger.info(`Successfully sent? %o`, info);
+        }
+      }
+    );
+    return;
+  }
+
   async getOrCreateTestUser(username: string): Promise<User> {
     let user = await this.getRepository().findOne({ username });
     if (!user) {
@@ -178,8 +268,8 @@ export class AccountService extends BaseService {
     }
     // test user, set fake data so they can immediately join a game
     // FIXME: use run-time configuration / settings to determine what user properties to bypass (passedQuiz, isVerified, hasParticipated, etc)
-    user.dateConsented = new Date();
-    user.isVerified = true;
+    // user.dateConsented = new Date();
+    // user.isVerified = true;
     // user.passedQuiz = true;
     await this.getRepository().save(user);
     return user;
